port remaining tests, delete json and json_tests modules
