+++ crates/ruff_db/src/diagnostic/render/json.rs
@@ -167,7 +167,24 @@ impl Serialize for ExpandedEdits<'_> {
 
 #[cfg(test)]
 mod tests {
-    use crate::diagnostic::{DiagnosticFormat, render::tests::create_notebook_diagnostics};
+    use crate::diagnostic::{
+        DiagnosticFormat,
+        render::tests::{
+            create_diagnostics, create_notebook_diagnostics, create_syntax_error_diagnostics,
+        },
+    };
+
+    #[test]
+    fn output() {
+        let (env, diagnostics) = create_diagnostics(DiagnosticFormat::Json);
+        insta::assert_snapshot!(env.render_diagnostics(&diagnostics));
+    }
+
+    #[test]
+    fn syntax_errors() {
+        let (env, diagnostics) = create_syntax_error_diagnostics(DiagnosticFormat::Json);
+        insta::assert_snapshot!(env.render_diagnostics(&diagnostics));
+    }
 
     #[test]
     fn notebook_output() {
