+++ crates/ruff_linter/src/message/mod.rs
@@ -14,8 +14,6 @@ use ruff_db::files::File;
 pub use github::GithubEmitter;
 pub use gitlab::GitlabEmitter;
 pub use grouped::GroupedEmitter;
-pub use json::JsonEmitter;
-pub use json_lines::JsonLinesEmitter;
 pub use junit::JunitEmitter;
 pub use pylint::PylintEmitter;
 pub use rdjson::RdjsonEmitter;
@@ -33,8 +31,6 @@ mod diff;
 mod github;
 mod gitlab;
 mod grouped;
-mod json;
-mod json_lines;
 mod junit;
 mod pylint;
 mod rdjson;
